requirements.txt
Flask>=2.2.0
Werkzeug>=2.2.0
SQLAlchemy>=1.4
Adafruit-Blinka
Adafruit-CircuitPython-Fingerprint
pyserial
folium
psycopg2
gunicorn

